+++ include/pgm/pgm_index_in_memory.hpp
@@ -44,7 +44,7 @@ namespace pgm {
  * @tparam TopLevelBitSize the bit-size of the cells in the top-level table, must be either 0 or a power of two
  * @tparam Floating the floating-point type to use for slopes
  */
-template<typename K, size_t Epsilon = 64, uint8_t TopLevelBitSize = 32, typename Floating = double>
+template<typename K, size_t Epsilon = 64, uint8_t TopLevelBitSize = 32, typename Floating = float>
 class InMemoryPGMIndex {
 protected:
     static_assert(Epsilon > 0);
